test(torghut): split exchange submission policy tests
